% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_spec.R
\name{new_cluster_spec}
\alias{new_cluster_spec}
\title{Functions required for celery-adjacent packages}
\usage{
new_cluster_spec(cls, args, eng_args, mode, method, engine)
}
\description{
These functions are helpful when creating new packages that will register
new cluster specifications.
}
\keyword{internal}
