% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_rules.R
\name{assoc_rules}
\alias{assoc_rules}
\title{Association Rules Mining}
\usage{
assoc_rules(
  mode = "partition",
  engine = "arules",
  method = "apriori",
  min_support = NULL,
  min_confidence = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model. The only
possible value for this model is "association".}

\item{engine}{A single character string specifying the computational engine
to use for fitting. The default for this model is \code{"arules"}.}

\item{method}{A single character string specifying the algorithm to use for
fitting. Possible algorithms are \code{"apriori"} and \code{"eclat"}. The default for
this model is \code{"apriori"}.}

\item{min_support}{Positive double, minimum support for a rule (between 0 and 1).}

\item{min_confidence}{Positive double, minimum confidence for a rule (between 0 and 1).}
}
\value{
A \code{assoc_rules} association specification.
}
\description{
\code{assoc_rules()} defines a model that finds association rules based on
specified minimum support and confidence.

The method of estimation is chosen by setting the model engine. The
engine-specific pages for this model are listed below.
\itemize{
\item \link[=details_assoc_rules_stats]{arules}
}
}
\details{
\subsection{What does it mean to predict?}{

WORK IN PROGRESS
}
}
\examples{
# Show all engines
modelenv::get_from_env("assoc_rules")

assoc_rules()
}
