% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_diagnostics.R
\name{avg_silhouette}
\alias{avg_silhouette}
\title{Measures average silhouette between clusters}
\usage{
avg_silhouette(.data, clusters, distance = "euclidean")
}
\arguments{
\item{.data}{A data frame}

\item{clusters}{The column containing cluster assignments.}

\item{distance}{The distance metric to use. (Currently only "euclidean")}
}
\value{
The average of all cluster silhouettes.
}
\description{
Measures average silhouette between clusters
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_celery("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

mtcars \%>\%
  augment(kmeans_fit) \%>\%
  silhouettes(.cluster)

}
