% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butcher.R
\name{axe-cluster_fit}
\alias{axe-cluster_fit}
\alias{axe_call.cluster_fit}
\alias{axe_ctrl.cluster_fit}
\alias{axe_data.cluster_fit}
\alias{axe_env.cluster_fit}
\alias{axe_fitted.cluster_fit}
\title{Axing a cluster_fit.}
\usage{
axe_call.cluster_fit(x, verbose = FALSE, ...)

axe_ctrl.cluster_fit(x, verbose = FALSE, ...)

axe_data.cluster_fit(x, verbose = FALSE, ...)

axe_env.cluster_fit(x, verbose = FALSE, ...)

axe_fitted.cluster_fit(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed cluster_fit object.
}
\description{
cluster_fit objects are created from the tidyclust package.
}
\examples{
\dontshow{if (rlang::is_installed("butcher")) withAutoprint(\{ # examplesIf}
k_fit <- k_means(num_clusters = 3) |>
  parsnip::set_engine("stats") |>
  fit(~., data = mtcars)

butcher::butcher(k_fit)
\dontshow{\}) # examplesIf}
}
