% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celery-package.R
\docType{package}
\name{celery-package}
\alias{celery}
\alias{celery-package}
\title{celery: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EmilHvitfeldt/celery}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/celery/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Authors:
\itemize{
  \item Kelly Bodwin \email{kelly@bodwin.us}
}

}
\keyword{internal}
