% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_empty_ellipse_celery}
\alias{check_empty_ellipse_celery}
\title{Check to ensure that ellipses are empty}
\usage{
check_empty_ellipse_celery(...)
}
\arguments{
\item{...}{Extra arguments.}
}
\value{
If an error is not thrown (from non-empty ellipses), a NULL list.
}
\description{
Check to ensure that ellipses are empty
}
\keyword{internal}
