% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data.R
\name{contr_one_hot}
\alias{contr_one_hot}
\title{One-hot contrast matrix}
\usage{
contr_one_hot(n, contrasts = TRUE, sparse = FALSE)
}
\arguments{
\item{n}{A vector of character factor levels (of length >=1) or the number
of unique levels (>= 1).}

\item{contrasts}{This argument is for backwards compatibility and only the
default of \code{TRUE} is supported.}

\item{sparse}{This argument is for backwards compatibility and only the
default of \code{FALSE} is supported.}
}
\description{
A re-export of \code{\link[hardhat:contr_one_hot]{hardhat::contr_one_hot()}} for use with
\code{indicators = "one_hot"}.
}
