% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_clust_hdbscan.R
\name{details_db_clust_hdbscan}
\alias{details_db_clust_hdbscan}
\title{Hierarchical Density-Based Spatial Clustering (HDBSCAN) via dbscan}
\description{
\code{\link[=db_clust]{db_clust()}} creates an HDBSCAN model.
}
\details{
For this engine, there is a single mode: partition
\subsection{Tuning Parameters}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Error in `dplyr::mutate()`:
## ℹ In argument: `dials = purrr::map(call_info, get_dials)`.
## Caused by error in `purrr::map()`:
## ℹ In index: 1.
## Caused by error:
## ! 'min_points' is not an exported object from 'namespace:dials'
}\if{html}{\out{</div>}}
}
}
\keyword{internal}
