% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_itemsets_arules.R
\name{details_freq_itemsets_arules}
\alias{details_freq_itemsets_arules}
\title{Frequent Itemsets via arules}
\description{
\code{\link[=freq_itemsets]{freq_itemsets()}} creates K-means model. This engine uses the classical definition
of a K-means model, which only takes numeric predictors.
}
\keyword{internal}
