% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_clust_mclust.R
\name{details_gm_clust_mclust}
\alias{details_gm_clust_mclust}
\title{Gaussian Mixture Model (GMM) via mclust}
\description{
\code{\link[=gm_clust]{gm_clust()}} creates GMM model.
}
\details{
For this engine, there is a single mode: partition
\subsection{Tuning Parameters}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Error in `dplyr::mutate()`:
## ℹ In argument: `dials = purrr::map(call_info, get_dials)`.
## Caused by error in `purrr::map()`:
## ℹ In index: 3.
## Caused by error:
## ! 'circular' is not an exported object from 'namespace:dials'
}\if{html}{\out{</div>}}
}
}
\keyword{internal}
