% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_rules.R
\name{.assoc_rules_fit_arules}
\alias{.assoc_rules_fit_arules}
\title{Simple Wrapper around arules functions}
\usage{
.assoc_rules_fit_arules(
  x,
  min_support = NULL,
  min_confidence = NULL,
  method = "apriori"
)
}
\arguments{
\item{min_support}{Minimum support threshold.}

\item{min_confidence}{Minimum confidence threshold.}

\item{method}{Algorithm to use for mining frequent itemsets. Either "apriori" or "eclat".}

\item{data}{A transaction data set.}
}
\value{
A set of association rules based on the specified parameters.
}
\description{
This wrapper prepares the data and parameters to send to either \code{arules::apriori}
or \code{arules::eclat} for association rules mining, depending on the chosen method.
}
\keyword{internal}
