% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_itemsets.R
\name{.freq_itemsets_fit_arules}
\alias{.freq_itemsets_fit_arules}
\title{Simple Wrapper around arules functions}
\usage{
.freq_itemsets_fit_arules(x, min_support = NULL, mining_method = NULL)
}
\arguments{
\item{x}{A transaction data set.}

\item{min_support}{Minimum support threshold.}

\item{mining_method}{Algorithm to use for mining frequent itemsets. Either "apriori" or "eclat".}
}
\value{
A set of frequent itemsets based on the specified parameters.
}
\description{
This wrapper prepares the data and parameters to send to either \code{arules::apriori}
or \code{arules::eclat} for frequent itemsets mining, depending on the chosen method.
}
\keyword{internal}
