% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_clust.R
\name{.gm_clust_fit_mclust}
\alias{.gm_clust_fit_mclust}
\title{Simple Wrapper around Mclust function}
\usage{
.gm_clust_fit_mclust(x, G = NULL, ...)
}
\arguments{
\item{x}{matrix or data frame}

\item{num_clusters}{Number of clusters}
}
\value{
mclust object
}
\description{
This wrapper prepares the data into a distance matrix to send to
\code{mclust::Mclust} and retains the parameters \code{num_clusters} as an
attribute.
}
\keyword{internal}
