% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{.hier_clust_fit_stats}
\alias{.hier_clust_fit_stats}
\title{Simple Wrapper around hclust function}
\usage{
.hier_clust_fit_stats(
  x,
  num_clusters = NULL,
  cut_height = NULL,
  linkage_method = NULL,
  dist_fun = philentropy::distance
)
}
\arguments{
\item{x}{matrix or data frame}

\item{num_clusters}{the number of clusters}

\item{cut_height}{the height to cut the dendrogram}

\item{linkage_method}{the agglomeration method to be used. This should be (an
unambiguous abbreviation of) one of \code{"ward.D"}, \code{"ward.D2"}, \code{"single"},
\code{"complete"}, \code{"average"} (= UPGMA), \code{"mcquitty"} (= WPGMA), \code{"median"} (=
WPGMC) or \code{"centroid"} (= UPGMC).}

\item{dist_fun}{A function of the form \verb{function(x)} that takes a data
frame or matrix and returns a \code{dist} object. Defaults to
\code{philentropy::distance} with Euclidean distance. See
\code{philentropy::getDistMethods()} for a list of supported methods, and
\code{vignette("tuning_and_metrics", package = "tidyclust")} for usage
examples.}
}
\value{
A dendrogram
}
\description{
This wrapper prepares the data into a distance matrix to send to
\code{stats::hclust} and retains the parameters \code{num_clusters} or \code{h} as an
attribute.
}
\keyword{internal}
