% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{.k_means_fit_clustMixType}
\alias{.k_means_fit_clustMixType}
\title{Simple Wrapper around clustMixType kmeans}
\usage{
.k_means_fit_clustMixType(x, k, ...)
}
\arguments{
\item{x}{Data frame with both numerics and factors (also ordered factors are possible).}

\item{k}{Either the number of clusters, a vector specifying indices of initial prototypes, or a data frame of 
prototypes of the same columns as \code{x}.}

\item{...}{Other arguments passed to \code{clustMixType::kproto()}}
}
\value{
Result from \code{clustMixType::kproto()}
}
\description{
This wrapper runs \code{clustMixType::kproto()} and reorders the clusters.
}
\keyword{internal}
