% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{.k_means_fit_klaR}
\alias{.k_means_fit_klaR}
\title{Simple Wrapper around klaR kmeans}
\usage{
.k_means_fit_klaR(data, modes, ...)
}
\arguments{
\item{data}{A matrix or data frame of categorical data. Objects have to be in rows, variables in columns.}

\item{modes}{Either the number of modes or a set of initial
    (distinct) cluster modes.  If a number, a random set of (distinct)
    rows in \code{data} is chosen as the initial modes.}

\item{...}{Other arguments passed to \code{klaR::kmodes()}}
}
\value{
Result from \code{klaR::kmodes()}
}
\description{
This wrapper runs \code{klaR::kmodes()} and reorders the clusters.
}
\keyword{internal}
