% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_shift.R
\name{.mean_shift_fit_LPCM}
\alias{.mean_shift_fit_LPCM}
\title{Simple Wrapper around LPCM::ms function}
\usage{
.mean_shift_fit_LPCM(x, bandwidth = NULL, ...)
}
\arguments{
\item{x}{matrix or data frame.}

\item{bandwidth}{Kernel bandwidth controlling the neighborhood size.}
}
\value{
ms object
}
\description{
This wrapper passes the data and bandwidth to \code{LPCM::ms()} with plotting
disabled.
}
\keyword{internal}
