% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{get_tidyclust_colors}
\alias{get_tidyclust_colors}
\title{Get colors for tidyclust text.}
\usage{
get_tidyclust_colors()
}
\value{
a list of \code{cli} functions.
}
\description{
Get colors for tidyclust text.
}
\keyword{internal}
