% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_metrics.R
\name{enrichment}
\alias{enrichment}
\title{Measures relationship between cluster assignments and another categorical variable.}
\usage{
enrichment(data, clusters, var)
}
\arguments{
\item{data}{the dataset}

\item{clusters}{the variable with cluster assignments}

\item{var}{other variables for enrichment}
}
\value{
The p-value of a Chi-Square test for relationship between cluster
assignments and the categorical variable.
}
\description{
Measures relationship between cluster assignments and another categorical variable.
}
