% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_assignment.R
\name{extract_cluster_assignment}
\alias{extract_cluster_assignment}
\title{Extract cluster assignments from model}
\usage{
extract_cluster_assignment(object, ...)
}
\arguments{
\item{object}{An cluster_spec object.}

\item{...}{Other arguments passed to methods}
}
\description{
Extract cluster assignments from model
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
set_engine_celery("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

kmeans_fit \%>\%
  predict(new_data = mtcars)

kmeans_fit \%>\%
  extract_cluster_assignment()

}
