% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cluster.R
\name{extract_clusters}
\alias{extract_clusters}
\title{Extract clusters from model}
\usage{
extract_clusters(object, ...)
}
\arguments{
\item{object}{An cluster_spec object.}

\item{...}{Other arguments passed to methods.}
}
\description{
Extract clusters from model
}
\examples{
set.seed(1234)
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_celery("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

kmeans_fit \%>\%
  extract_clusters()
}
