% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_predictions.R
\name{extract_predictions}
\alias{extract_predictions}
\title{Returns recommender predictions with predicted values imputed into dataset
Notes: currently imputes thresholded probabilities}
\usage{
extract_predictions(pred_output)
}
\description{
Returns recommender predictions with predicted values imputed into dataset
Notes: currently imputes thresholded probabilities
}
