% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{finalize_model_tidyclust}
\alias{finalize_model_tidyclust}
\alias{finalize_workflow_tidyclust}
\title{Splice final parameters into objects}
\usage{
finalize_model_tidyclust(x, parameters)

finalize_workflow_tidyclust(x, parameters)
}
\arguments{
\item{x}{A recipe, \code{parsnip} model specification, or workflow.}

\item{parameters}{A list or 1-row tibble of parameter values. Note that the
column names of the tibble should be the \code{id} fields attached to \code{tune()}.
For example, in the \code{Examples} section below, the model has \code{tune("K")}. In
this case, the parameter tibble should be "K" and not "neighbors".}
}
\value{
An updated version of \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions are deprecated. Please use \code{\link[tune:finalize_model]{tune::finalize_model()}} and
\code{\link[tune:finalize_model]{tune::finalize_workflow()}} instead, which now support \code{cluster_spec}
objects natively.
}
\examples{
kmeans_spec <- k_means(num_clusters = tune())
best_params <- data.frame(num_clusters = 5)

# Old:
finalize_model_tidyclust(kmeans_spec, best_params)

# New:
tune::finalize_model(kmeans_spec, best_params)
}
