% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_x}
\alias{fit_x}
\title{Estimate model parameters.}
\usage{
fit_x(object, ...)
}
\arguments{
\item{object}{An cluster_spec object.}

\item{...}{Other arguments passed to methods}
}
\description{
Estimates parameters for a given model from a set of data in the form of
a set of predictors (\code{x}).
}
