% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_models.R
\name{get_model_env_celery}
\alias{get_model_env_celery}
\alias{get_from_env_celery}
\alias{set_env_val_celery}
\title{Working with the celery model environment}
\usage{
get_model_env_celery()

get_from_env_celery(items)

set_env_val_celery(name, value)
}
\arguments{
\item{items}{A character string of objects in the model environment.}

\item{name}{A single character value for a new symbol in the model environment.}

\item{value}{A single value for a new value in the model environment.}

\item{...}{Named values that will be assigned to the model environment.}
}
\description{
These functions read and write to the environment where the package stores
information about model specifications.
}
\examples{
# Access the model data:
current_code <- get_model_env_celery()
ls(envir = current_code)
}
\keyword{internal}
