% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-params.R
\name{gm_clust_params}
\alias{gm_clust_params}
\alias{circular}
\alias{zero_covariance}
\alias{shared_orientation}
\alias{shared_shape}
\alias{shared_size}
\title{Gaussian mixture covariance structure parameters}
\usage{
circular(values = c(TRUE, FALSE))

zero_covariance(values = c(TRUE, FALSE))

shared_orientation(values = c(TRUE, FALSE))

shared_shape(values = c(TRUE, FALSE))

shared_size(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{A vector of possible values (\code{c(TRUE, FALSE)} by default).}
}
\value{
A \code{dials} parameter object for use with \code{\link[tune:tune_grid]{tune::tune_grid()}} and
related functions.
}
\description{
Logical flags controlling the covariance structure of cluster Gaussians
fit by \code{tidyclust::gm_clust()} with the \code{mclust} engine. See
\code{\link[=gm_clust]{gm_clust()}} for descriptions.
}
\examples{
circular()
zero_covariance()
shared_orientation()
shared_shape()
shared_size()
}
