% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{gmm}
\alias{gmm}
\title{Gaussian Mixture Model (GMM) Clustering}
\usage{
gmm(mode = "partition", engine = "ClusterR", num_clusters = NULL)
}
\arguments{
\item{mode}{A single character string for the type of model. The only
possible value for this model is "partition".}

\item{engine}{A single character string specifying what computational engine
to use for fitting. Possible engines are listed below. The default for this
model is \code{"ClusterR"}.}

\item{num_clusters}{Positive integer, number of clusters in model.}
}
\value{
A \code{gmm} cluster specification.
}
\description{
\code{gmm()} defines a model that fits clusters based on distances to a number
of centers.
}
\examples{
# Show all engines
modelenv::get_from_env("gmm")

gmm()
}
