% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-jaccard.R
\name{jaccard_avg}
\alias{jaccard_avg}
\alias{jaccard_avg.cluster_spec}
\alias{jaccard_avg.cluster_fit}
\alias{jaccard_avg.workflow}
\title{Estimate the Average Jaccard Index for Cluster Stability}
\usage{
jaccard_avg(object, ...)

\method{jaccard_avg}{cluster_spec}(object, ...)

\method{jaccard_avg}{cluster_fit}(object, new_data = NULL, ...)

\method{jaccard_avg}{workflow}(object, new_data = NULL, ...)
}
\arguments{
\item{object}{A fitted tidyclust model}

\item{...}{Not currently used.}

\item{new_data}{A dataset to predict on.}
}
\value{
A tibble with 3 columns; \code{.metric}, \code{.estimator}, and \code{.estimate}.
}
\description{
This function estimates the stability of cluster assignments by calculating
the average Jaccard index: how well do cluster assignments of new data from a
pre-trained model align with cluster assignments of the same data from a new
model fitted on the same data (Hennig, 2007).
}
\details{
Note that this metric will re-fit the model on the new dataset
(\code{new_data}), and so might be computationally expensive.
}
\examples{
library(tidymodels)

penguins <- drop_na(penguins)

kmeans_spec <- k_means(num_clusters = tune())

res <- tune_cluster(
  kmeans_spec,
  ~ bill_length_mm + bill_depth_mm + flipper_length_mm + body_mass_g,
  resamples = bootstraps(penguins, times = 10),
  metrics = cluster_metric_set(jaccard_avg)
)

# What's the largest k that has high recovery rate?
# (values above 0.75 are considered good, values below 0.5 are considered bad)
autoplot(res)

}
\references{
Hennig, C. (2007). Cluster-wise assessment of cluster stability.
\emph{Computational Statistics & Data Analysis, 52}(1), 258-271.
}
\seealso{
\code{\link[fpc:clusterboot]{fpc::clusterboot()}}

Other cluster metric: 
\code{\link{pred_strength}()},
\code{\link{silhouette_avg}()},
\code{\link{sse_ratio}()},
\code{\link{sse_total}()},
\code{\link{sse_within_total}()}
}
\concept{cluster metric}
