% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials.R
\name{k}
\alias{k}
\title{Number of Clusters}
\usage{
k(range = c(1L, 10L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::log10_trans()} or \code{scales::reciprocal_trans()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Number of Clusters
}
\examples{
k()
}
