% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{K-Means}
\usage{
k_means(mode = "partition", engine = "stats", num_clusters = NULL)
}
\arguments{
\item{mode}{A single character string for the type of model. The only
possible value for this model is "partition".}

\item{engine}{A single character string specifying what computational engine
to use for fitting. Possible engines are listed below. The default for this
model is \code{"stats"}.}

\item{num_clusters}{Positive integer, number of clusters in model.}
}
\description{
\code{k_means()} defines a model that fits clusters based on distances to a number
of centers.
}
\examples{
# Show all engines
modelenv::get_from_env("k_means")

k_means()
}
