% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-params.R
\docType{data}
\name{linkage_method}
\alias{linkage_method}
\alias{values_linkage_method}
\title{The agglomeration Linkage method}
\format{
An object of class \code{character} of length 8.
}
\usage{
linkage_method(values = values_linkage_method)

values_linkage_method
}
\arguments{
\item{values}{A character string of possible values. See \code{linkage_methods}
in examples below.}
}
\description{
The agglomeration Linkage method
}
\details{
This parameter is used in \code{tidyclust} models for \code{hier_clust()}.
}
\examples{
values_linkage_method
linkage_method()
}
\keyword{datasets}
