% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_spec.R
\name{make_classes_celery}
\alias{make_classes_celery}
\title{Prepend a new class}
\usage{
make_classes_celery(prefix)
}
\arguments{
\item{prefix}{A character string for a class.}
}
\value{
A character vector.
}
\description{
This adds an extra class to a base class of "cluster_spec".
}
\keyword{internal}
