% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_clust.R
\name{mclust_helper}
\alias{mclust_helper}
\title{mclust fit helper function}
\usage{
mclust_helper(
  circular,
  zero_covariance,
  shared_orientation,
  shared_shape,
  shared_size
)
}
\arguments{
\item{object}{gm_clust object}
}
\value{
string containing mclust model name
}
\description{
This function returns the mclust model name based on the specified
TRUE/FALSE model arguments
}
\keyword{internal}
