% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_helpers.R
\name{min_grid.cluster_spec}
\alias{min_grid.cluster_spec}
\title{Determine the minimum set of model fits}
\usage{
\method{min_grid}{cluster_spec}(x, grid, ...)
}
\arguments{
\item{x}{A cluster specification.}

\item{grid}{A tibble with tuning parameter combinations.}

\item{...}{Not currently used.}
}
\value{
A tibble with the minimum tuning parameters to fit and an additional
list column with the parameter combinations used for prediction.
}
\description{
Determine the minimum set of model fits
}
