% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dials-params.R
\name{min_points}
\alias{min_points}
\title{Minimum number of points}
\usage{
min_points(range = c(2L, 20L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\value{
A \code{dials} parameter object for use with \code{\link[tune:tune_grid]{tune::tune_grid()}} and
related functions.
}
\description{
The minimum number of connected points required to form a core point in
density-based clustering. Used in \code{tidyclust::db_clust()} with the \code{dbscan}
and \code{hdbscan} engines.
}
\examples{
min_points()
}
