% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_spec.R
\name{new_cluster_spec}
\alias{new_cluster_spec}
\title{Functions required for tidyclust-adjacent packages}
\usage{
new_cluster_spec(cls, args, eng_args, mode, method, engine)
}
\arguments{
\item{cls}{A single character string for the model type (e.g. \code{"k_means"}).}

\item{args}{A named list of main model arguments.}

\item{eng_args}{A named list of engine-specific arguments.}

\item{mode}{A single character string for the model mode (e.g.
\code{"partition"}).}

\item{method}{A list of method details or \code{NULL}.}

\item{engine}{A single character string for the computational engine.}
}
\value{
A \code{\link{cluster_spec}} object made to work with tidyclust.
}
\description{
These functions are helpful when creating new packages that will register new
cluster specifications.
}
\keyword{internal}
