% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cluster.R
\name{predict_cluster}
\alias{predict_cluster}
\alias{predict_cluster.cluster_fit}
\title{Other predict methods.}
\usage{
predict_cluster(object, ...)

\method{predict_cluster}{cluster_fit}(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{\link{cluster_fit}}.}

\item{...}{Optional arguments passed to the underlying predict function.
Use \code{prefix} to change the prefix in the cluster factor levels (default:
\code{"Cluster_"}). Use \code{labels} to supply a character vector of cluster labels,
which overrides \code{prefix}.}

\item{new_data}{A rectangular data object, such as a data frame.}
}
\value{
A \code{tibble::tibble()}.

A \code{tibble::tibble()}.
}
\description{
These are internal functions not meant to be directly called by the user.
}
\keyword{internal}
