% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-pred_strength.R
\name{pred_strength}
\alias{pred_strength}
\alias{pred_strength.cluster_spec}
\alias{pred_strength.cluster_fit}
\alias{pred_strength.workflow}
\title{Estimate the Prediction Strength of a Clustering Model}
\usage{
pred_strength(object, ...)

\method{pred_strength}{cluster_spec}(object, ...)

\method{pred_strength}{cluster_fit}(object, new_data = NULL, ...)

\method{pred_strength}{workflow}(object, new_data = NULL, ...)
}
\arguments{
\item{object}{A fitted tidyclust model}

\item{...}{Not currently used.}

\item{new_data}{A dataset to predict on.}
}
\value{
A tibble with 3 columns; \code{.metric}, \code{.estimator}, and \code{.estimate}.
}
\description{
This function estimates the stability of cluster assignments by calculating
the prediction strength: how well do cluster assignments of new data from a
pre-trained model co-incide with cluster assignments of the same data from a
new model fitted on the same data (Tibshirani & Walther, 2005).
}
\details{
Note that this metric will re-fit the model on the new dataset
(\code{new_data}), and so might be computationally expensive.
}
\examples{

library(tidymodels)

penguins <- drop_na(penguins)

kmeans_spec <- k_means(num_clusters = tune())

res <- tune_cluster(
  kmeans_spec,
  ~ bill_length_mm + bill_depth_mm + flipper_length_mm + body_mass_g,
  resamples = vfold_cv(penguins, v = 2, repeats = 10),
  metrics = cluster_metric_set(pred_strength)
)

# What's the largest k that has a good prediction strength?
# (values above 0.8 are considered good)
autoplot(res)

}
\references{
Tibshirani, R. and Walther, G. (2005) Cluster Validation by
Prediction Strength, \emph{Journal of Computational and Graphical Statistics, 14}(3), 511-528.
}
\seealso{
Other cluster metric: 
\code{\link{jaccard_avg}()},
\code{\link{silhouette_avg}()},
\code{\link{sse_ratio}()},
\code{\link{sse_total}()},
\code{\link{sse_within_total}()}
}
\concept{cluster metric}
