% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cluster_fit}
\alias{print.cluster_fit}
\alias{print.cluster_spec}
\title{Print a cluster object}
\usage{
\method{print}{cluster_fit}(x, ...)

\method{print}{cluster_spec}(x, ...)
}
\arguments{
\item{x}{A \code{\link{cluster_fit}} or \code{\link{cluster_spec}} object.}

\item{...}{Arguments passed to the underlying print method.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Print a cluster object
}
\keyword{internal}
