% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_pkgs.R
\name{required_pkgs.cluster_spec}
\alias{required_pkgs.cluster_spec}
\alias{required_pkgs.cluster_fit}
\title{Get required packages for a cluster object}
\usage{
\method{required_pkgs}{cluster_spec}(x, infra = TRUE, ...)

\method{required_pkgs}{cluster_fit}(x, infra = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link{cluster_spec}} or \code{\link{cluster_fit}} object.}

\item{infra}{A logical. Should tidyclust itself be included in the result?}

\item{...}{Currently unused.}
}
\value{
A character vector of required package names.
}
\description{
Get required packages for a cluster object
}
\keyword{internal}
