% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{set_args.cluster_spec}
\alias{set_args.cluster_spec}
\title{Change arguments of a cluster specification}
\usage{
\method{set_args}{cluster_spec}(object, ...)
}
\arguments{
\item{object}{A model specification.}

\item{...}{One or more named model arguments.}
}
\value{
An updated \code{\link{cluster_spec}} object.
}
\description{
Change arguments of a cluster specification
}
