% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{set_args_tidyclust}
\alias{set_args_tidyclust}
\alias{set_mode_tidyclust}
\title{Change elements of a cluster specification}
\usage{
set_args_tidyclust(object, ...)

set_mode_tidyclust(object, mode)
}
\arguments{
\item{object}{A cluster specification.}

\item{...}{One or more named model arguments.}

\item{mode}{A character string for the model type (e.g. "classification" or
"regression")}
}
\value{
An updated model object.
}
\description{
\code{set_args_tidyclust()} can be used to modify the arguments of a cluster
specification while \code{set_mode()} is used to change the model's mode.
}
\details{
\code{set_args_tidyclust()} will replace existing values of the
arguments.
}
\examples{
k_means()

k_means() \%>\%
  set_args_tidyclust(k = 3) \%>\%
  set_mode_tidyclust("partition")

}
