% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{set_engine.cluster_spec}
\alias{set_engine.cluster_spec}
\title{Change engine of a cluster specification}
\usage{
\method{set_engine}{cluster_spec}(object, engine, ...)
}
\arguments{
\item{object}{A model specification.}

\item{engine}{A character string for the software that should
be used to fit the model. This is highly dependent on the type
of model (e.g. linear regression, random forest, etc.).}

\item{...}{Any optional arguments associated with the chosen computational
engine. These are captured as quosures and can be tuned with \code{tune()}.}
}
\value{
An updated \code{\link{cluster_spec}} object.
}
\description{
Change engine of a cluster specification
}
