% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{set_engine_celery}
\alias{set_engine_celery}
\title{Declare a computational engine and specific arguments}
\usage{
set_engine_celery(object, engine, ...)
}
\arguments{
\item{object}{A model specification.}

\item{engine}{A character string for the software that should
be used to fit the model. This is highly dependent on the type
of model (e.g. K-means, etc.).}

\item{...}{Any optional arguments associated with the chosen computational
engine. These are captured as quosures and can be \code{tune()}.}
}
\value{
An updated model specification.
}
\description{
\code{set_engine_celery()} is used to specify which package or system will be used
to fit the model, along with any arguments specific to that software.
}
\section{Engines}{

Based on the currently loaded packages, the following lists the set of
engines available to each model specification.
}

\examples{
# First, set general arguments using the standardized names
mod <- k_means(k = 10) \%>\%
  # now say how you want to fit the model and another other options
  set_engine_celery("stats", iter.max = 15)

translate_celery(mod, engine = "stats")
}
