% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{set_mode.cluster_spec}
\alias{set_mode.cluster_spec}
\title{Change mode of a cluster specification}
\usage{
\method{set_mode}{cluster_spec}(object, mode, ...)
}
\arguments{
\item{object}{A model specification.}

\item{mode}{A character string for the model type (e.g. "classification" or
"regression")}

\item{...}{One or more named model arguments.}
}
\value{
An updated \code{\link{cluster_spec}} object.
}
\description{
Change mode of a cluster specification
}
