% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_diagnostics.R
\name{silhouettes}
\alias{silhouettes}
\title{Measures silhouettes between clusters}
\usage{
silhouettes(.dist, clusters)
}
\arguments{
\item{.dist}{A distance matrix}

\item{clusters}{A vector containing cluster assignments in the
row order of the distance matrix.}
}
\value{
The silhouettes matrix.
}
\description{
Measures silhouettes between clusters
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_celery("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

dists <- mtcars \%>\%
  as.matrix() \%>\%
  dist()

silhouettes(dists, kmeans_fit$fit$cluster)

}
