% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_diagnostics.R
\name{sse_ratio}
\alias{sse_ratio}
\title{Compute the ratio of the WSS to the total SSE}
\usage{
sse_ratio(object, ...)
}
\arguments{
\item{object}{An cluster_spec object.}

\item{...}{Other arguments passed to methods.}
}
\description{
Compute the ratio of the WSS to the total SSE
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_celery("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

kmeans_fit \%>\%
  sse_ratio()
}
