% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{stats_kmeans_fit}
\alias{stats_kmeans_fit}
\title{Simple Wrapper around stats kmeans}
\usage{
stats_kmeans_fit(data, centers = NULL, ...)
}
\arguments{
\item{centers}{either the number of clusters, say \eqn{k}, or a set of
    initial (distinct) cluster centres.  If a number, a random set of
    (distinct) rows in \code{x} is chosen as the initial centres.}

\item{...}{Other arguments passed to \code{stats::kmeans()}}
}
\value{
Result from \code{stats::kmeans()}
}
\description{
This wrapper runs \code{stats::kmeans} and adds a check that \code{centers} is
specified
}
\keyword{internal}
