% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butcher.R
\name{tidyclust-butcher}
\alias{tidyclust-butcher}
\alias{axe_env.k_means}
\alias{axe_env.gmm}
\alias{axe_env.hier_clust}
\alias{axe_env.cluster_fit}
\alias{axe_env.workflow}
\title{Butcher methods for tidyclust}
\usage{
axe_env.k_means(x, verbose = FALSE, ...)

axe_env.gmm(x, verbose = FALSE, ...)

axe_env.hier_clust(x, verbose = FALSE, ...)

axe_env.cluster_fit(x, verbose = FALSE, ...)

axe_env.workflow(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A tidyclust cluster object.}

\item{verbose}{Should information be printed about how much memory is freed
from butchering?}

\item{...}{Extra arguments possibly used by underlying methods.}
}
\description{
These methods allow you to use the butcher package to reduce the size of
a tidyclust cluster object. After calling \code{butcher::butcher()} on an object,
the only guarantee is that you will still be able to \code{predict()} from that
object. Other functions may not work as expected.
}
