% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_diagnostics.R
\name{tot_wss}
\alias{tot_wss}
\title{Compute the sum of within-cluster SSE}
\usage{
tot_wss(object, ...)
}
\arguments{
\item{object}{An cluster_spec object.}

\item{...}{Other arguments passed to methods.}
}
\description{
Compute the sum of within-cluster SSE
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_celery("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

kmeans_fit \%>\%
  tot_wss()
}
