% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{tunable.cluster_spec}
\alias{tunable.cluster_spec}
\alias{tunable.hier_clust}
\alias{tunable.k_means}
\alias{tunable.db_clust}
\alias{tunable.gm_clust}
\title{Get tunable parameters for a cluster specification}
\usage{
\method{tunable}{cluster_spec}(x, ...)

\method{tunable}{hier_clust}(x, ...)

\method{tunable}{k_means}(x, ...)

\method{tunable}{db_clust}(x, ...)

\method{tunable}{gm_clust}(x, ...)
}
\arguments{
\item{x}{An object, such as a recipe, recipe step, workflow, or model
specification.}

\item{...}{Other arguments passed to methods}
}
\value{
A tibble with columns \code{name}, \code{call_info}, \code{source}, \code{component},
and \code{component_id} describing each tunable parameter.
}
\description{
Get tunable parameters for a cluster specification
}
\keyword{internal}
