% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_args.R
\name{tune_args.cluster_spec}
\alias{tune_args.cluster_spec}
\title{Get tune arguments for a cluster specification}
\usage{
\method{tune_args}{cluster_spec}(object, full = FALSE, ...)
}
\arguments{
\item{object}{A \code{model_spec}, \code{recipe}, \code{workflow}, or other object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A tibble describing the tunable arguments in the cluster
specification.
}
\description{
Get tune arguments for a cluster specification
}
\keyword{internal}
