% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-sse.R
\name{within_cluster_sse}
\alias{within_cluster_sse}
\title{Calculates Sum of Squared Error in each cluster}
\usage{
within_cluster_sse(object, new_data = NULL, dist_fun = Rfast::dista)
}
\arguments{
\item{object}{A fitted kmeans tidyclust model}

\item{new_data}{A dataset to predict on.  If \code{NULL}, uses trained clustering.}

\item{dist_fun}{A function for calculating distances to centroids.  Defaults
to Euclidean distance on processed data.}
}
\value{
A tibble with two columns, the cluster name and the SSE within that
cluster.
}
\description{
Calculates Sum of Squared Error in each cluster
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_tidyclust("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

kmeans_fit \%>\%
  within_cluster_sse()

}
