% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_itemset_predict.R
\name{augment_itemset_predict}
\alias{augment_itemset_predict}
\title{Extract Predictions from Observation Data Frames}
\usage{
augment_itemset_predict(pred_output, truth_output)
}
\arguments{
\item{pred_output}{A data frame with one column, where each cell contains a data frame.}
}
\value{
A data frame with items as columns and non-NA values as rows.
}
\description{
This function processes a data frame containing observation data frames and extracts non-NA values.
}
\details{
Returns recommender predictions with predicted values imputed into dataset
Notes: currently imputes thresholded probabilities
}
