% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-silhouette.R
\name{avg_silhouette}
\alias{avg_silhouette}
\alias{avg_silhouette.cluster_fit}
\alias{avg_silhouette_vec}
\title{Measures average silhouette across all observations}
\usage{
avg_silhouette(object, ...)

\method{avg_silhouette}{cluster_fit}(object, new_data = NULL, dists = NULL, dist_fun = NULL, ...)

avg_silhouette_vec(
  object,
  new_data = NULL,
  dists = NULL,
  dist_fun = Rfast::Dist,
  ...
)
}
\arguments{
\item{object}{A fitted kmeans tidyclust model}

\item{...}{Other arguments passed to methods.}

\item{new_data}{A dataset to predict on.  If \code{NULL}, uses trained clustering.}

\item{dists}{A distance matrix. Used if \code{new_data} is \code{NULL}.}

\item{dist_fun}{A function for calculating distances between observations.
Defaults to Euclidean distance on processed data.}
}
\value{
A double; the average silhouette.
}
\description{
Measures average silhouette across all observations
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_tidyclust("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

dists <- mtcars \%>\%
  as.matrix() \%>\%
  dist()

avg_silhouette(kmeans_fit, dists = dists)

avg_silhouette_vec(kmeans_fit, dists = dists)
}
