% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-aaa.R
\name{cluster_metric_set}
\alias{cluster_metric_set}
\title{Combine metric functions}
\usage{
cluster_metric_set(...)
}
\arguments{
\item{...}{The bare names of the functions to be included in the metric set.
These functions must be cluster metrics such as \code{\link[=sse_total]{sse_total()}},
\code{\link[=sse_ratio]{sse_ratio()}}, or \code{\link[=silhouette_avg]{silhouette_avg()}}.}
}
\value{
A \code{cluster_metric_set()} object, combining the use of all input
metrics.
}
\description{
\code{cluster_metric_set()} allows you to combine multiple metric functions
together into a new function that calculates all of them at once.
}
\details{
All functions must be:
\itemize{
\item Only cluster metrics
}
}
