% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_clust.R
\name{db_clust}
\alias{db_clust}
\title{Density-Based Spatial Clustering of Applications with Noise (DBSCAN)}
\usage{
db_clust(
  mode = "partition",
  engine = "dbscan",
  radius = NULL,
  min_points = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model. The only
possible value for this model is "partition".}

\item{engine}{A single character string specifying what computational engine
to use for fitting. The engine for this model is \code{"dbscan"}.}

\item{radius}{Positive integer, Radius used to determine core-points and cluster points together (required).}

\item{min_points}{Positive double, Minimum number of points needed to form a cluster (required)}
}
\value{
A \code{db_clust} cluster specification.
}
\description{
\code{db_clust} defines a model that fits clusters based on areas with observations
that are densely packed together

There are different ways to fit this model, and the method of estimation is
chosen by setting the model engine. The engine-specific pages for this model
are listed below.
\itemize{
\item \link[=details_db_clust_dbscan]{dbscan}
}
}
\details{
\subsection{What does it mean to predict?}{

To predict the cluster assignment for a new observation, we determine if a point
is within the radius of a core point. If so, we predict the same cluster as the core point.
If not, we predict the observation to be an outlier.
}
}
\examples{
# Show all engines
modelenv::get_from_env("db_clust")

db_clust()
}
