% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_clust.R
\name{dbscan_helper}
\alias{dbscan_helper}
\title{dbscan fit helper function}
\usage{
dbscan_helper(object, ...)
}
\arguments{
\item{x}{matrix or data frame}
}
\value{
dbscan object
}
\description{
This function...
}
\keyword{internal}
