% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_clust.R
\name{.db_clust_fit_dbscan}
\alias{.db_clust_fit_dbscan}
\title{Simple Wrapper around dbscan function}
\usage{
.db_clust_fit_dbscan(x, eps = NULL, minPts = NULL, ...)
}
\arguments{
\item{x}{matrix or data frame}

\item{radius}{Radius used to determine core-points and cluster points together}

\item{minpts}{Minimum number of points needed to form a cluster}
}
\value{
dbscan object
}
\description{
This wrapper prepares the data into a distance matrix to send to
\code{dbscan::dbscan} and retains the parameters \code{radius} or \code{minpts} as an
attribute.
}
\keyword{internal}
