% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_clust.R
\name{.gm_clust_fit_mclust}
\alias{.gm_clust_fit_mclust}
\title{Simple Wrapper around Mclust function}
\usage{
.gm_clust_fit_mclust(
  x,
  num_clusters = NULL,
  circular = NULL,
  zero_covariance = NULL,
  shared_orientation = NULL,
  shared_shape = NULL,
  shared_size = NULL,
  ...
)
}
\arguments{
\item{x}{matrix or data frame.}

\item{num_clusters}{Number of clusters.}

\item{circular}{Whether or not to fit circular MVG distributions for each cluster.}

\item{zero_covariance}{Whether or not to assign covariances of 0 for each MVG.}

\item{shared_orientation}{Whether each cluster MVG should have the same orientation.}

\item{shared_shape}{Whether each cluster MVG should have the same shape.}

\item{shared_size}{Whether each cluster MVG should have the same size/volume.}
}
\value{
mclust object
}
\description{
This wrapper prepares the data into a distance matrix to send to
\code{mclust::Mclust} and retains the parameters \code{num_clusters} as an
attribute.
}
\keyword{internal}
