% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{.k_means_fit_clustMixType}
\alias{.k_means_fit_clustMixType}
\title{Simple Wrapper around clustMixType kmeans}
\usage{
.k_means_fit_clustMixType(x, k, ...)
}
\arguments{
\item{...}{Other arguments passed to \code{clustMixType::kproto()}}
}
\value{
Result from \code{clustMixType::kproto()}
}
\description{
This wrapper runs \code{clustMixType::kproto()} and reorders the clusters.
}
\keyword{internal}
