% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_predictions.R
\name{extract_predictions}
\alias{extract_predictions}
\title{Extract Predictions from Observation Data Frames}
\usage{
extract_predictions(pred_output)
}
\arguments{
\item{pred_output}{A data frame with one column, where each cell contains a data frame.}
}
\value{
A data frame with items as columns and non-NA values as rows.
}
\description{
This function processes a data frame containing observation data frames and extracts non-NA values.
}
