% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-helpers.R
\name{get_centroid_dists}
\alias{get_centroid_dists}
\title{Computes distance from observations to centroids}
\usage{
get_centroid_dists(
  new_data,
  centroids,
  dist_fun = function(x, y) {
     philentropy::dist_many_many(x, y, method =
    "euclidean")
 }
)
}
\arguments{
\item{new_data}{A data frame}

\item{centroids}{A data frame where each row is a centroid.}

\item{dist_fun}{A function for computing matrix-to-matrix distances. Defaults
to
\code{function(x, y) philentropy::dist_many_many(x, y, method = "euclidean")}.}
}
\description{
Computes distance from observations to centroids
}
