% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{glance.cluster_fit}
\alias{glance.cluster_fit}
\title{Construct a single row summary "glance" of a model, fit, or other object}
\usage{
\method{glance}{cluster_fit}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\value{
a tibble
}
\description{
This method glances the model in a tidyclust model object, if it exists.
}
