% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_clust.R
\name{mclust_helper}
\alias{mclust_helper}
\title{mclust fit helper function}
\usage{
mclust_helper(
  circular,
  zero_covariance,
  shared_orientation,
  shared_shape,
  shared_size
)
}
\arguments{
\item{circular}{Whether or not to fit circular MVG distributions for each cluster.}

\item{zero_covariance}{Whether or not to assign covariances of 0 for each MVG.}

\item{shared_orientation}{Whether each cluster MVG should have the same orientation.}

\item{shared_shape}{Whether each cluster MVG should have the same shape.}

\item{shared_size}{Whether each cluster MVG should have the same size/volume.}
}
\value{
string containing mclust model name
}
\description{
This function returns the mclust model name based on the specified
TRUE/FALSE model arguments.
}
\keyword{internal}
