% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-aaa.R
\name{new_cluster_metric}
\alias{new_cluster_metric}
\title{Construct a new clustering metric function}
\usage{
new_cluster_metric(fn, direction)
}
\arguments{
\item{fn}{A function.}

\item{direction}{A string. One of:
\itemize{
\item \code{"maximize"}
\item \code{"minimize"}
\item \code{"zero"}
}}
}
\value{
A \code{cluster_metric} object.
}
\description{
These functions provide convenient wrappers to create the one
type of metric functions in celrry: clustering metrics. They add a
metric-specific class to \code{fn}. These features are used by
\code{\link[=cluster_metric_set]{cluster_metric_set()}} and by \code{\link[=tune_cluster]{tune_cluster()}} when tuning.
}
