% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cluster.R
\name{predict_cluster}
\alias{predict_cluster}
\alias{predict_cluster.cluster_fit}
\title{Other predict methods.}
\usage{
predict_cluster(object, ...)

\method{predict_cluster}{cluster_fit}(object, new_data, ...)
}
\arguments{
\item{object}{An object of class \code{cluster_fit}}

\item{...}{Arguments to the underlying model's prediction
function cannot be passed here (see \code{opts}).}

\item{new_data}{A rectangular data object, such as a data frame.}
}
\description{
These are internal functions not meant to be directly called by the user.
}
\keyword{internal}
