% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-helpers.R
\name{prep_data_dist}
\alias{prep_data_dist}
\title{Prepares data and distance matrices for metric calculation}
\usage{
prep_data_dist(
  object,
  new_data = NULL,
  dists = NULL,
  dist_fun = philentropy::distance
)
}
\arguments{
\item{object}{A fitted \code{\link{cluster_spec}} object.}

\item{new_data}{A dataset to calculate predictions on.  If \code{NULL}, the
trained cluster assignments from the fitted object are used.}

\item{dists}{A distance matrix for the data.  If \code{NULL}, distance is computed
on \code{new_data} using the \code{stats::dist()} function.}

\item{dist_fun}{A custom distance functions.}
}
\value{
A list
}
\description{
Prepares data and distance matrices for metric calculation
}
