% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_diagnostics.R
\name{silhouettes}
\alias{silhouettes}
\title{Measures silhouettes between clusters}
\usage{
silhouettes(.data, clusters, distance = "euclidean")
}
\arguments{
\item{.data}{A data frame}

\item{clusters}{The column containing cluster assignments.}

\item{distance}{The distance metric to use. (Currently only "euclidean")}
}
\value{
The silhouettes matrix.
}
\description{
Measures silhouettes between clusters
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_celery("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

mtcars \%>\%
  augment(kmeans_fit) \%>\%
  silhouettes(.cluster)

}
