% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-silhouette.R
\name{silhouettes}
\alias{silhouettes}
\title{Measures silhouettes between clusters}
\usage{
silhouettes(object, new_data = NULL, dists = NULL, dist_fun = Rfast::Dist)
}
\arguments{
\item{object}{A fitted kmeans tidyclust model}

\item{new_data}{A dataset to predict on.  If \code{NULL}, uses trained clustering.}

\item{dists}{A distance matrix. Used if \code{new_data} is \code{NULL}.}

\item{dist_fun}{A function for calculating distances between observations.
Defaults to Euclidean distance on processed data.}
}
\value{
A tibble giving the silhouettes for each observation.
}
\description{
Measures silhouettes between clusters
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_tidyclust("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

dists <- mtcars \%>\%
  as.matrix() \%>\%
  dist()

silhouettes(kmeans_fit, dists = dists)
}
