% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_metrics.R
\name{sse_ratio}
\alias{sse_ratio}
\title{Compute the ratio of the WSS to the total SSE}
\usage{
sse_ratio(object, new_data = NULL, dist_fun = Rfast::dista, ...)
}
\arguments{
\item{object}{A fitted kmeans celery model}

\item{new_data}{A dataset to predict on.  If \code{NULL}, uses trained clustering.}

\item{dist_fun}{A function for calculating distances to centroids.  Defaults
to Euclidean distance on processed data.}

\item{...}{Other arguments passed to methods.}
}
\description{
Compute the ratio of the WSS to the total SSE
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_celery("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

kmeans_fit \%>\%
  sse_ratio()
}
