% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric-sse.R
\name{sse_within}
\alias{sse_within}
\alias{sse_within.cluster_fit}
\alias{sse_within.workflow}
\alias{sse_within_vec}
\title{Compute the sum of within-cluster SSE}
\usage{
sse_within(object, ...)

\method{sse_within}{cluster_fit}(object, new_data = NULL, dist_fun = NULL, ...)

\method{sse_within}{workflow}(object, new_data = NULL, dist_fun = NULL, ...)

sse_within_vec(object, new_data = NULL, dist_fun = Rfast::dista, ...)
}
\arguments{
\item{object}{A fitted kmeans tidyclust model}

\item{...}{Other arguments passed to methods.}

\item{new_data}{A dataset to predict on.  If \code{NULL}, uses trained clustering.}

\item{dist_fun}{A function for calculating distances to centroids.  Defaults
to Euclidean distance on processed data.}
}
\description{
Compute the sum of within-cluster SSE
}
\examples{
kmeans_spec <- k_means(num_clusters = 5) \%>\%
  set_engine("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

sse_within(kmeans_fit)

sse_within_vec(kmeans_fit)
}
