% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.cluster_fit}
\alias{tidy.cluster_fit}
\title{Turn a celery model object into a tidy tibble}
\usage{
\method{tidy}{cluster_fit}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{...}{Additional arguments to tidying method.}
}
\value{
a tibble
}
\description{
This method tidies the model in a celery model object, if it exists.
}
