% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyclust-package.R
\docType{package}
\name{tidyclust-package}
\alias{tidyclust}
\alias{tidyclust-package}
\title{tidyclust: A Common API to Clustering}
\description{
A common interface to specifying clustering models, in the same style as 'parsnip'. Creates unified interface across different functions and computational engines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/tidyclust}
  \item \url{https://tidyclust.tidymodels.org/}
  \item Report bugs at \url{https://github.com/tidymodels/tidyclust/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emil.hvitfeldt@posit.co} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Authors:
\itemize{
  \item Kelly Bodwin \email{kelly@bodwin.us}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (03wc8by49) [copyright holder, funder]
}

}
\keyword{internal}
