% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_clust.R, R/gm_clust.R, R/hier_clust.R,
%   R/k_means.R, R/mean_shift.R, R/update.R
\name{update.db_clust}
\alias{update.db_clust}
\alias{update.gm_clust}
\alias{update.hier_clust}
\alias{update.k_means}
\alias{update.mean_shift}
\alias{tidyclust_update}
\title{Update a cluster specification}
\usage{
\method{update}{db_clust}(
  object,
  parameters = NULL,
  radius = NULL,
  min_points = NULL,
  fresh = FALSE,
  ...
)

\method{update}{gm_clust}(
  object,
  parameters = NULL,
  num_clusters = NULL,
  circular = NULL,
  zero_covariance = NULL,
  shared_orientation = NULL,
  shared_shape = NULL,
  shared_size = NULL,
  fresh = FALSE,
  ...
)

\method{update}{hier_clust}(
  object,
  parameters = NULL,
  num_clusters = NULL,
  cut_height = NULL,
  linkage_method = NULL,
  dist_fun = NULL,
  fresh = FALSE,
  ...
)

\method{update}{k_means}(object, parameters = NULL, num_clusters = NULL, fresh = FALSE, ...)

\method{update}{mean_shift}(object, parameters = NULL, bandwidth = NULL, fresh = FALSE, ...)
}
\arguments{
\item{object}{A cluster specification.}

\item{parameters}{A 1-row tibble or named list with \emph{main} parameters to
update. Use \strong{either} \code{parameters} \strong{or} the main arguments directly when
updating. If the main arguments are used, these will supersede the values
in \code{parameters}. Also, using engine arguments in this object will result in
an error.}

\item{radius}{Positive double, Radius drawn around points to determine core-points and cluster assignments (required).}

\item{min_points}{Positive integer, Minimum number of connected points required to form a core-point, including the point itself (required).}

\item{fresh}{A logical for whether the arguments should be modified in-place
or replaced wholesale.}

\item{...}{Not used for \code{update()}.}

\item{num_clusters}{Positive integer, number of clusters in model.}

\item{circular}{Boolean, whether or not to fit circular MVG distributions for each cluster. Default \code{TRUE}.}

\item{zero_covariance}{Boolean, whether or not to assign covariances of 0 for each MVG. Default \code{TRUE}.}

\item{shared_orientation}{Boolean, whether each cluster MVG should have the same orientation. Default \code{TRUE}.}

\item{shared_shape}{Boolean, whether each cluster MVG should have the same shape. Default \code{TRUE}.}

\item{shared_size}{Boolean, whether each cluster MVG should have the same size/volume. Default \code{TRUE}.}

\item{cut_height}{Positive double, height at which to cut dendrogram to
obtain cluster assignments (only used if \code{num_clusters} is \code{NULL})}

\item{linkage_method}{the agglomeration method to be used. This should be (an
unambiguous abbreviation of) one of \code{"ward.D"}, \code{"ward.D2"}, \code{"single"},
\code{"complete"}, \code{"average"} (= UPGMA), \code{"mcquitty"} (= WPGMA), \code{"median"} (=
WPGMC) or \code{"centroid"} (= UPGMC).}

\item{dist_fun}{A function for calculating the distance between observations.
Defaults to \code{philentropy::distance} which supports numerous distance
metrics via its \code{method} argument. The function should accept a matrix or
data frame and return a square numeric matrix or an object coercible to one
via \code{\link[stats:dist]{stats::as.dist()}}. See \code{\link[=silhouette]{silhouette()}} for further details.}
}
\value{
An updated cluster specification.
}
\description{
If parameters of a cluster specification need to be modified,
\code{update()} can be used in lieu of recreating the object from scratch.
}
\examples{
kmeans_spec <- k_means(num_clusters = 5)
kmeans_spec
update(kmeans_spec, num_clusters = 1)
update(kmeans_spec, num_clusters = 1, fresh = TRUE)

param_values <- tibble::tibble(num_clusters = 10)

kmeans_spec |> update(param_values)
}
