% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_metrics.R
\name{tot_sse}
\alias{tot_sse}
\title{Compute the total sum of squares}
\usage{
tot_sse(object, new_data = NULL, dist_fun = Rfast::dista, ...)
}
\arguments{
\item{object}{A fitted kmeans tidyclust model}

\item{new_data}{A dataset to predict on.  If \code{NULL}, uses trained clustering.}

\item{dist_fun}{A function for calculating distances to centroids.  Defaults
to Euclidean distance on processed data.}

\item{...}{Other arguments passed to methods.}
}
\description{
Compute the total sum of squares
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_tidyclust("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

kmeans_fit \%>\%
  tot_sse()
}
