% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_diagnostics.R
\name{within_cluster_sse}
\alias{within_cluster_sse}
\title{Calculates Sum of Squared Error in each cluster}
\usage{
within_cluster_sse(object, ...)
}
\arguments{
\item{object}{a fitted kmeans celery model}

\item{...}{Other arguments passed to methods.}
}
\value{
A tibble with two columns, the cluster name and the SSE within that
cluster.
}
\description{
Calculates Sum of Squared Error in each cluster
}
\examples{
kmeans_spec <- k_means(k = 5) \%>\%
  set_engine_celery("stats")

kmeans_fit <- fit(kmeans_spec, ~., mtcars)

kmeans_fit \%>\%
  within_cluster_sse()

}
